backend/ivy-gcp-installer/requirements.txt
google-cloud-secret-manager>=2.16.0
google-cloud-logging>=3.6.0
google-cloud-monitoring>=2.15.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-gcp-trace>=1.4.0
pyyaml>=6.0
requests>=2.31.0
pydantic>=2.4.0
